fix tsoa: use autogenerated routes, fully benefit from decorators, introduce UUID class, fix get user with id

+++ apiserver/src/controllers/UserController.ts
@@ -1,3 +1,4 @@
+import { UUID } from './../types/express/UUID';
 import {
 	Body,
 	Controller,
@@ -11,24 +12,23 @@ import {
 	SuccessResponse,
 	Tags,
 } from "@tsoa/runtime";
-import UserService from "../services/UserService";
-import User from "../models/User";
-import MailAlreadyUsedError from "../errors/UserErrors/MailAlreadyUsedError";
-import MailIsInvalidError from "../errors/UserErrors/MailIsInvalidError";
-import FieldIsMissingError from "../errors/FieldIsMissingError";
-import UserNotFoundError from "../errors/UserErrors/UserNotFoundError";
+import UserService from "./../services/UserService";
+import User from "./../models/User";
+import MailAlreadyUsedError from "./../errors/UserErrors/MailAlreadyUsedError";
+import MailIsInvalidError from "./../errors/UserErrors/MailIsInvalidError";
+import FieldIsMissingError from "./../errors/FieldIsMissingError";
+import UserNotFoundError from "./../errors/UserErrors/UserNotFoundError";
+import NotUUIDError from '../errors/NotUUIDError';
+import { isValidEmail, isValidUUID } from "./../helpers/validators";
 
 type CreateUserDTO = Omit<User, "id" | "friends" | "createdDate">;
 type UserIdDTO = Pick<User, "id">;
 type UserDTO = Pick<User, "displayName" | "email">;
 
-const EMAIL_REGEX =
-	/^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
-
-@Security("bearerAuth") // Follow https://github.com/lukeautry/tsoa/issues/1082 for root-level security
+@Security("auth0") // Follow https://github.com/lukeautry/tsoa/issues/1082 for root-level security
 @Route("users")
 @Tags("User")
-class UserController extends Controller {
+export class UserController extends Controller {
 	/**
 	 * Create a new user during sign up. Even if users are authenticated and
 	 * created by Auth0, we manage a user database to store preferences,
@@ -37,13 +37,15 @@ class UserController extends Controller {
 	 * @returns {Promise<UserDTO>} the created user
 	 */
 	@SuccessResponse(200)
-	@Response<Error>(400)
-	@Put("/")
-	static async create(@Body() body: CreateUserDTO): Promise<UserIdDTO> {
+	@Response<FieldIsMissingError>(400)
+	@Response<MailIsInvalidError>(400)
+	@Response<MailAlreadyUsedError>(400)
+	@Put("")
+	async create(@Body() body: CreateUserDTO): Promise<UserIdDTO> {
 		if (!body.email || !body.displayName) {
 			throw new FieldIsMissingError(!body.email ? "email" : "displayName");
 		}
-		if (!EMAIL_REGEX.test(body.email)) {
+		if (!isValidEmail(body.email)) {
 			throw new MailIsInvalidError();
 		}
 		try {
@@ -59,10 +61,13 @@ class UserController extends Controller {
 
 	/**
 	 * Delete logged user.
-	 * @param {string} userID the GUID of user
+	 * @param {UUID} userID the GUID of user
 	 */
-	@Delete("/{userId}")
-	static async delete(@Path() userId: string): Promise<void> {
+	@Delete("{userId}")
+	async delete(@Path() userId: UUID): Promise<void> {
+		if (!isValidUUID(userId)) {
+			throw new NotUUIDError(userId);
+		}
 		await UserService.delete(userId);
 	}
 
@@ -70,8 +75,8 @@ class UserController extends Controller {
 	 * Gets all user's data.
 	 * @returns {Promise<User>} the user
 	 */
-	@Get("/")
-	static async getAll(): Promise<UserDTO[]> {
+	@Get()
+	async getAll(): Promise<UserDTO[]> {
 		const users: User[] = await UserService.getAll();
 		return users.map((user) => {
 			const { id, friends, createdDate, ...rest } = user;
@@ -81,18 +86,20 @@ class UserController extends Controller {
 
 	/**
 	 * Gets logged user's data.
-	 * @param {string} userID the GUID of user
-	 * @returns {Promise<User>} the user
+	 * @param {UUID} userID the GUID of user
+	 * @returns {Promise<UserDTO>} the user
 	 */
-	@Get("/{userId}")
-	static async get(@Path() userId: string): Promise<UserDTO> {
+	@Get("{userId}")
+	async get(@Path() userId: UUID): Promise<UserDTO> {
+		if (!isValidUUID(userId)) {
+			throw new NotUUIDError(userId);
+		}
 		const user: User | undefined = await UserService.get(userId);
 		if (!user) {
 			throw new UserNotFoundError();
 		} else {
-			return user;
+			const { id, createdDate, ...rest} = user;
+			return rest;
 		}
 	}
 }
-
-export default UserController;
